backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "cohere>=5.20.0",
    "fastapi>=0.124.0",
    "openai-agents>=0.6.2",
    "qdrant-client>=1.7.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.38.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "beautifulsoup4>=4.12.0",
]

[tool.setuptools.packages.find]
where = ["src"]
